Add characteristic enemy and  levels 1-5

+++ Assets/Source/Griny/Scripts/Attacker.cs
@@ -5,9 +5,9 @@ using UnityEngine.Serialization;
 public abstract class Attacker: MonoBehaviour
 {
     [SerializeField] private Color _damageColor;
-    [SerializeField] private int _damage;
+    [SerializeField] private float _damage;
 
-    public int Damage => _damage;
+    public float Damage => _damage;
     public Color DamageColor => _damageColor;
 
     private void OnEnable()
